Agregando log

+++ app/util/Main.js
@@ -119,7 +119,7 @@ function onMail(mail) {
                                 var code = getCountryCode(mail.subject);
                                 PreciManager.getPreciByCode(code, function (credit) {
                                     console.log()
-                                    if (parseInt(client.credit) >= parseInt(credit)) {
+                                    if (parseFloat(client.credit) >= parseFloat(credit)) {
                                         ClientManager.updateClientCredit(client, credit);
                                         RequestManager.createRequest('Request', client, credit);
                                         SmsManager.send(client.phone, mail.text);
